pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yamlguard"
version = "0.1.0"
description = "Fast, accurate CLI for YAML indentation and Kubernetes manifest validation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "YAMLGuard Team", email = "team@yamlguard.dev"}
]
keywords = ["yaml", "kubernetes", "linting", "validation", "indentation", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "ruamel.yaml>=0.18.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "jsonschema>=4.19.0",
    "tomli>=2.0.0",
    "tomli-w>=1.0.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "urllib3>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "hypothesis>=6.82.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]
secrets = [
    "detect-secrets>=1.4.0",
]
kube = [
    "kubernetes>=28.0.0",
]

[project.scripts]
yamlguard = "yamlguard.cli:app"

[project.urls]
Homepage = "https://github.com/yamlguard/yamlguard"
Repository = "https://github.com/yamlguard/yamlguard"
Documentation = "https://yamlguard.readthedocs.io"
"Bug Tracker" = "https://github.com/yamlguard/yamlguard/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["yamlguard*"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["yamlguard"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=yamlguard",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "kube: marks tests that require Kubernetes schemas",
    "secrets: marks tests that scan for secrets",
]

[tool.coverage.run]
source = ["yamlguard"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/.*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
